+++ src/main/java/org/skriptlang/skript/bukkit/input/InputKey.java
@@ -1,10 +1,14 @@
-package org.skriptlang.skript.bukkit;
+package org.skriptlang.skript.bukkit.input;
 
 import org.bukkit.Input;
 
 import java.util.EnumSet;
 import java.util.Set;
 
+/**
+ * Enum representing different movement input keys.
+ * @see Input
+ */
 public enum InputKey {
 
 	FORWARD("forward movement key"),
@@ -21,6 +25,12 @@ public enum InputKey {
 		this.toString = toString;
 	}
 
+	/**
+	 * Checks if the given {@link Input} is pressing this {@link InputKey}.
+	 *
+	 * @param input the input to check
+	 * @return true if the {@link Input} is pressing this {@link InputKey}, false otherwise
+	 */
 	public boolean check(Input input) {
 		return switch (this) {
 			case FORWARD -> input.isForward();
@@ -38,6 +48,12 @@ public enum InputKey {
 		return toString;
 	}
 
+	/**
+	 * Returns a set of {@link InputKey}s that match the given {@link Input}.
+	 *
+	 * @param input the input to check
+	 * @return a set of {@link InputKey}s that match the given {@link Input}
+	 */
 	public static Set<InputKey> fromInput(Input input) {
 		Set<InputKey> keys = EnumSet.noneOf(InputKey.class);
 		for (InputKey key : values()) {
